+++ app/src/main/java/com/example/gagyeboost/model/data/DateItem.kt
@@ -1,4 +1,4 @@
-package com.example.gagyeboost.ui.home
+package com.example.gagyeboost.model.data
 
 data class DateItem(
     val expense: Int?,
